--- Objt/Entry.cs
@@ -1,6 +1,0 @@
-﻿namespace API_Pessoa.Objt
-{
-    public class Entry
-    {
-    }
-}
