+++ README.md
@@ -6,6 +6,7 @@ Criar cad_Pais						->
 Criar cad_Unidade_Federativa		-> 
 Criar cad_Municipio					-> 
 
+Criar cad_Cliente					-> FEITO!
 Criar cad_Companhia_Transporte
 
 
@@ -13,6 +14,10 @@ Criar cad_Companhia_Transporte
 GET ALL cad_Pessoa					-> FEITO!
 GET por ID cad_Pessoa				-> FEITO!
 Insert no cad_Pessoa				-> 
+
+GET ALL cad_Cliente					-> FEITO!
+GET por ID cad_Cliente				-> FEITO!
+Insert no cad_Cliente				-> 
 
 Insert Cliente
 Insert Exportador
