+++ cad_Cliente/Objt/Entry.cs
@@ -1,0 +1,6 @@
+﻿namespace API_Pessoa.cad_Cliente.Objt
+{
+    public class Entry
+    {
+    }
+}
